Nixify emacs
This begins the process of nixifying these dotfiles. First up is
emacs.

Emacs is now configured to respect external package manager's
packages, although it will still install them with use-package if need
be.

This also contains a nix expression for an emacs with all
use-package'd packages, though this is still a little imperfect.

[Bauer](https://github.com/matthewbauer/bauer) was a huge inspiration
for this, thanks!