Nixify stumpwm
This completely refactors the way stumpwm is set up; instead of using
quicklisp in the config file, we instead rely on the system asdf being
set up correctly. This should be more robust (since we can hook into
the system package manager, which will hopefully recompile things when
the sbcl binary changes).

We also change the stumpwm config so that it can be reloaded using
`(loadrc)`, and hook that functionality into home-manager, so that
reloading configuration is more convenient.

This also moves xsession management to stumpwm, and sets up a more
reasonable stumpwm package.

Note that we need to hack around package dependencies - all
dependencies and their reverse dependencies must be explicitly set in
`~/.config/common-lisp/source-registry.conf`. This is annoying,
especially since this list is already created in
`$NIX_LISP_ASDF_PATHS`, just not in any kind of useful format. I
should open an upstream issue.