stumpwm: Use a user cache to avoid writing to nix store fasl files
Previously, since the stumpwm modules repository was written to the
read-only nix store, loading any modules from there would break
because asdf would attempt to create object files in the source tree,
since nix configures it to expect object files in the source
tree (yes... This isn't as awful if the modules are pre-compiled).

This made it impossible to use the nixos wrapper directly. Instead, we
now set up custom caching in stumpwm, overriding the old behavior, and
we get a much nicer configuration.