dunst: Move config into nix
This is necessary becuase the window geometry is decided by screen
size, but dunst itself doesn't support dividing the screen width.

So we move it into nix, which can (with the help of a little variable)
work out the correct window size for the computer the configuration is
installed to.