emacs: Switch to the upstream-provided user service
The only thing the local one did was define an ExecReload, but this
was buggy anyway. If we use the upstream user service we can
automatically create an `emacsclient.desktop` file as well, which is
more useful than a buggy reload.

The latter is still quite cool, if I ever figure out how to make it
work I'll just push for it upstream.