Build emacs from unstable to avoid issues with csv-mode
For some reason, the upstream sha of csv-mode changed. This results
in build failures for the nix package, since the source can no longer
be found.

For the time being, I'll just build emacs with unstable. This seems to
work fine, and side-steps this issue; when 21.5 lands I can presumably
move back.