emacs: Fix init files not being precompiled
It seems I didn't notice this for a long time because when installed
with a flaked NixOS module, home-manager will a) not display the error
without `-L` and b) just install the init files anyway.

This has no real effect on NixOS hosts, the files will just be
compiled next time emacs starts, but it can cause some odd behavior on
non-nix hosts (and show a bunch of annoying errors and warnings).