commit-nvfetcher: Don't create no-op commits
As it turns out, nvfetcher will update its database even if there are
no updates.

Obviously we don't want to create commits (and worse, PRs) in that
case, so we check if the changelog contains any updates before
committing.

We also clean up after ourselves if this happens, in case the script
is run manually.