Fix the emacs package override
The overlay was previously a bit botched. While it worked fine for
elpa packages (which was the primary use case), it didn't work for
melpa packages, because those rely on the `pkgs` used by
`emacsPackagesNgGen`.

This fixed the symptom it should have, but the override itself was
faulty. the `pkgs` provided to the local packages was just normal,
stable pkgs, with no overrides applied, so a normal emacs was built.

This fixes that (by making sure `prev` points to a patched nixpkgs),
and also takes the liberty of updating to the new name for
`emacsPackagesNgGen`.