Switch to eglot instead of lsp-mode
Mostly due to its insistence on downloading server binaries, and it
being quite impossible to dynamically load the PATH for its
implementing packages to adapt.

Originally I wanted flycheck for its `list-errors` integration, so I
couldn't us eglot, but flymake seems to support that now (and better!)
so eglot is a straight upgrade.

The upstream code for lsp-mode is also a bit of a huge horror fest,
which is counter the whole idea of language servers IMHO.

eglot is apparently even slanted for inclusion in upstream emacs, so
there's really no reason not to use it.