emacs: Refactor package and override project dependency
This was necessary because the project package, once again, updated
their tarball without releasing a new version (*vigorous fist
shaking*). Hacking it downstream until upstream can be updated.

Also figured I should refactor this while I'm here, so that if this
happens again in the future it'll be easier to figure out how to fix.