emacs: Implement auto-insert snippets without yatemplate
Yatemplate was cool, but it had the problematic side effect of
requiring data to be encoded in file names. Writing regexes in file
names isn't fun, so instead I'm setting the auto insert definitions
manually and explicitly now.

Eventually it would maybe be cool to implement this functionality as a
directive in yasnippet instead, but there isn't much upstream
maintenance for that package anymore and figuring out how to order
modes seems like a pain.