system-update: Work around git CVE fix
Don't *really* understand this CVE, it basically assumes full write
access to the user's system (at least at user permission level) from
another user, which means we can also modify `~/.config/git` and PWN
systems that way (or by any number of other dotfiles in the user's
home).

I guess it helps if you're not careful about the write permissions on
your home directory but are for `~/.config` somehow? But even then git
will fall back to `~/.git`, so you could create that directory... Kind
of don't understand how this fixes anything, it feels like doing
something to tick the "fixed CVE" checkmark.

It comes at a cost of breaking practically the whole planet's build
systems, mind you, we're going to see a lot of complaints about git
for the next year or so.