fix(nvidia): Use upstream powermanagement as this now works
I tried to debug #383 again, and discovered that the udev module now
just works as expected.

I don't know why, perhaps an upstream change in boot. That, or I
simply made a mistake while originally digging into this; if so, well,
I've lived under the belief that powermanagement wasn't working for
a *while* now, sorry about that!

Fixes #383